pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # Core stacks
  "transformers==4.46.3",
  "diffusers==0.31.0",
  "huggingface-hub==0.25.2",
  "accelerate==0.31.0",
  # Models & utils
  "sentence-transformers==5.1.1",
  "torch>=2.2",
  "torchvision>=0.17.0",
  "safetensors>=0.6.2",
  "pillow>=10.2",
  "numpy>=1.26",
  "pandas>=2.0",
  "timm>=1.0.20",
  "einops>=0.8.1",
  "pydantic>=2.12.0",
  "protobuf==4.25.3",
  "soundfile>=0.13.1",
  "sacremoses>=0.1.1",
  "sentencepiece>=0.2.1",
  "pytesseract>=0.3.13",
  "emoji>=2.15.0",
  "pyctcdecode>=0.5.0",
]
